Extract two classes, Grid and Printer
